backend/requirements/dev.txt
# This file was autogenerated by uv via the following command:
#    uv export --frozen --group dev --group test --no-hashes -o backend/requirements/dev.txt
annotated-types==0.7.0
    # via pydantic
asgiref==3.9.1
    # via
    #   django
    #   django-allauth
    #   django-cors-headers
    #   django-ninja-extra
bandit==1.9.4
boolean-py==5.0
    # via license-expression
cachecontrol==0.14.4
    # via pip-audit
certifi==2025.8.3
    # via requests
cffi==2.0.0
    # via
    #   backend
    #   cryptography
charset-normalizer==3.4.3
    # via requests
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   bandit
    #   pytest
    #   qrcode
contextlib2==21.6.0
    # via django-ninja-extra
coverage==7.13.5
cryptography==46.0.7
    # via
    #   backend
    #   fido2
    #   pyjwt
cyclonedx-python-lib==11.7.0
    # via pip-audit
defusedxml==0.7.1
    # via py-serializable
dj-database-url==3.0.1
    # via backend
django==5.2.13
    # via
    #   backend
    #   dj-database-url
    #   django-allauth
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-formtools
    #   django-ninja
    #   django-ninja-extra
    #   django-ninja-jwt
    #   django-otp
    #   django-phonenumber-field
    #   django-two-factor-auth
django-allauth==65.14.1
    # via backend
django-cors-headers==4.7.0
    # via backend
django-debug-toolbar==6.0.0
django-formtools==2.5.1
    # via django-two-factor-auth
django-ninja==1.4.3
    # via
    #   backend
    #   django-ninja-extra
django-ninja-extra==0.30.1
    # via
    #   backend
    #   django-ninja-jwt
django-ninja-jwt==5.3.7
    # via backend
django-otp==1.6.1
    # via
    #   backend
    #   django-two-factor-auth
django-phonenumber-field==8.1.0
    # via
    #   backend
    #   django-two-factor-auth
django-two-factor-auth==1.17.0
    # via backend
fido2==2.0.0
    # via django-allauth
filelock==3.29.0
    # via cachecontrol
googleapis-common-protos==1.74.0
    # via opentelemetry-exporter-otlp-proto-http
gunicorn==23.0.0
    # via backend
idna==3.10
    # via requests
importlib-metadata==8.7.1
    # via opentelemetry-api
iniconfig==2.3.0
    # via pytest
injector==0.22.0
    # via django-ninja-extra
license-expression==30.4.4
    # via cyclonedx-python-lib
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
msgpack==1.1.2
    # via cachecontrol
openfeature-sdk==0.9.0
    # via backend
opentelemetry-api==1.41.1
    # via
    #   backend
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-logging
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp-proto-common==1.41.1
    # via opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-http==1.41.1
    # via backend
opentelemetry-instrumentation==0.62b1
    # via
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-logging
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-wsgi
opentelemetry-instrumentation-django==0.62b1
    # via backend
opentelemetry-instrumentation-logging==0.62b1
    # via backend
opentelemetry-instrumentation-requests==0.62b1
    # via backend
opentelemetry-instrumentation-wsgi==0.62b1
    # via opentelemetry-instrumentation-django
opentelemetry-proto==1.41.1
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.41.1
    # via
    #   backend
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.62b1
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
opentelemetry-util-http==0.62b1
    # via
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-wsgi
packageurl-python==0.17.6
    # via cyclonedx-python-lib
packaging==25.0
    # via
    #   gunicorn
    #   opentelemetry-instrumentation
    #   pip-audit
    #   pip-requirements-parser
    #   pytest
parameterized==0.9.0
phonenumberslite==9.0.12
    # via django-phonenumber-field
pillow==12.2.0
    # via backend
pip==26.1
    # via pip-api
pip-api==0.0.34
    # via pip-audit
pip-audit==2.10.0
pip-requirements-parser==32.0.1
    # via pip-audit
platformdirs==4.9.6
    # via pip-audit
pluggy==1.6.0
    # via pytest
protobuf==6.33.6
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
psycopg==3.2.3
    # via backend
py-serializable==2.1.0
    # via cyclonedx-python-lib
pycparser==2.22 ; implementation_name != 'PyPy'
    # via cffi
pydantic==2.12.0
    # via
    #   backend
    #   django-ninja
pydantic-core==2.41.1
    # via pydantic
pygments==2.20.0
    # via
    #   pytest
    #   rich
pyjwt==2.12.1
    # via
    #   backend
    #   django-ninja-jwt
pyparsing==3.3.2
    # via pip-requirements-parser
pypng==0.20220715.0
    # via qrcode
pytest==9.0.3
    # via pytest-django
pytest-django==4.12.0
python-decouple==3.8
    # via backend
pyyaml==6.0.2
    # via
    #   bandit
    #   django-allauth
qrcode==7.4.2
    # via
    #   django-allauth
    #   django-two-factor-auth
requests==2.33.0
    # via
    #   backend
    #   cachecontrol
    #   opentelemetry-exporter-otlp-proto-http
    #   pip-audit
rich==15.0.0
    # via
    #   bandit
    #   pip-audit
ruff==0.12.10
sortedcontainers==2.4.0
    # via cyclonedx-python-lib
sqlparse==0.5.4
    # via
    #   django
    #   django-debug-toolbar
stevedore==5.7.0
    # via bandit
tomli==2.4.1
    # via pip-audit
tomli-w==1.2.0
    # via pip-audit
typing-extensions==4.15.0
    # via
    #   backend
    #   cyclonedx-python-lib
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   psycopg
    #   pydantic
    #   pydantic-core
    #   qrcode
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2 ; sys_platform == 'win32'
    # via
    #   django
    #   psycopg
urllib3==2.6.3
    # via requests
whitenoise==6.9.0
    # via backend
wrapt==2.1.2
    # via opentelemetry-instrumentation
zipp==3.23.1
    # via importlib-metadata
